Changed the transfer of the pointer to the array with timers to transfer the ref array
Changed the implementation of removing the timer from the list:

previously timers were first marked as deleted, and then when the time was changed they were deleted from the array, but they were deleted from the list immediately. Because of this, there was a mismatch between the timer index in the list and the timer index in the array. Now the timer is immediately removed from both the array and the list